+++ src/AceSorting.h
@@ -33,8 +33,8 @@ SOFTWARE.
  * Click on the "Files" menu above to see the list of header files.
  */
 
-#ifndef ACE_SEGMENT_ACE_SEGMENT_H
-#define ACE_SEGMENT_ACE_SEGMENT_H
+#ifndef ACE_SORTING_ACE_SORTING_H
+#define ACE_SORTING_ACE_SORTING_H
 
 // Blacklist platforms using https://github.com/arduino/ArduinoCore-api due to
 // incompatibilities.
@@ -43,8 +43,8 @@ SOFTWARE.
 #endif
 
 // Version format: xxyyzz == "xx.yy.zz"
-#define ACE_SORTING_VERSION 300
-#define ACE_SORTING_VERSION_STRING "0.3.0"
+#define ACE_SORTING_VERSION 10000
+#define ACE_SORTING_VERSION_STRING "1.0.0"
 
 #include "ace_sorting/swap.h"
 #include "ace_sorting/bubbleSort.h"